Merges ccfile to common based on redalert version.
Restores Load_Alloc_Data that takes a reference to tiberiandawn jshell.cpp for compatibility.
Merged mixfile.cpp as well which was heldback until ccfile unification.
